+++ Libplanet.Tests/Store/StoreTest.cs
@@ -390,10 +390,8 @@ namespace Libplanet.Tests.Store
             Assert.False(Fx.Store.ContainsBlock(Fx.Block3.Hash));
         }
 
-        [SkippableTheory]
-        [InlineData(true)]
-        [InlineData(false)]
-        public void TxExecution(bool actionsLogsList)
+        [SkippableFact]
+        public void TxExecution()
         {
             void AssertTxSuccessesEqual(TxSuccess expected, TxExecution actual)
             {
@@ -401,7 +399,6 @@ namespace Libplanet.Tests.Store
                 var success = (TxSuccess)actual;
                 Assert.Equal(expected.TxId, success.TxId);
                 Assert.Equal(expected.BlockHash, success.BlockHash);
-                Assert.Equal(expected.ActionsLogsList, success.ActionsLogsList);
                 Assert.Equal(expected.UpdatedStates, success.UpdatedStates);
                 Assert.Equal(expected.FungibleAssetsDelta, success.FungibleAssetsDelta);
                 Assert.Equal(expected.UpdatedFungibleAssets, success.UpdatedFungibleAssets);
@@ -413,7 +410,6 @@ namespace Libplanet.Tests.Store
                 var failure = (TxFailure)actual;
                 Assert.Equal(expected.TxId, failure.TxId);
                 Assert.Equal(expected.BlockHash, failure.BlockHash);
-                Assert.Equal(expected.ActionsLogsList, failure.ActionsLogsList);
                 Assert.Equal(expected.ExceptionName, failure.ExceptionName);
                 Assert.Equal(expected.ExceptionMetadata, failure.ExceptionMetadata);
             }
@@ -427,15 +423,6 @@ namespace Libplanet.Tests.Store
             var inputA = new TxSuccess(
                 Fx.Hash1,
                 Fx.TxId1,
-                actionsLogsList
-                    ? new List<IReadOnlyList<string>>
-                    {
-                        new List<string>
-                        {
-                            "LOG",
-                        },
-                    }
-                    : null,
                 ImmutableDictionary<Address, IValue>.Empty.Add(
                     random.NextAddress(),
                     (Text)"state value"
@@ -467,7 +454,6 @@ namespace Libplanet.Tests.Store
             var inputB = new TxFailure(
                 Fx.Hash1,
                 Fx.TxId2,
-                actionsLogsList ? new List<IReadOnlyList<string>>() : null,
                 "AnExceptionName",
                 Dictionary.Empty.Add("foo", 1).Add("bar", "baz")
             );
@@ -481,7 +467,6 @@ namespace Libplanet.Tests.Store
             var inputC = new TxFailure(
                 Fx.Hash2,
                 Fx.TxId1,
-                actionsLogsList ? new List<IReadOnlyList<string>>() : null,
                 "AnotherExceptionName",
                 null
             );