+++ Libplanet/Action/ActionEvaluation.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using Bencodex.Types;
 using Libplanet.State;
 
@@ -20,20 +19,17 @@ namespace Libplanet.Action
         /// evaluate <paramref name="action"/>.</param>
         /// <param name="outputState">The result states that
         /// <paramref name="action"/> makes.</param>
-        /// <param name="logs">The logs recorded while executing action.</param>
         /// <param name="exception">An exception that has risen during evaluating a given
         /// <paramref name="action"/>.</param>
         public ActionEvaluation(
             IAction action,
             IActionContext inputContext,
             IAccountStateDelta outputState,
-            Exception? exception = null,
-            IReadOnlyList<string>? logs = null)
+            Exception? exception = null)
         {
             Action = action;
             InputContext = inputContext;
             OutputState = outputState;
-            Logs = logs ?? new List<string>();
             Exception = exception;
         }
 
@@ -61,11 +57,5 @@ namespace Libplanet.Action
         /// An exception that had risen during evaluation.
         /// </summary>
         public Exception? Exception { get; }
-
-        /// <summary>
-        /// Logs recorded while executing an <see cref="IAction"/> through
-        /// <see cref="IActionContext.PutLog(string)"/>.
-        /// </summary>
-        public IReadOnlyList<string> Logs { get; }
     }
 }